% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUMAPPlotData.R
\name{makeUMAPPlotData}
\alias{makeUMAPPlotData}
\title{Make UMAP plot data}
\usage{
makeUMAPPlotData(umap_mat, labels)
}
\arguments{
\item{labels}{A vector of labels.}

\item{umap_layout}{The \code{layout} matrix within a \code{umap} object.}
}
\value{
A data.frame of the UMAP coordinates and the labels.
}
\description{
Using a given set of labels and umap output, create a data.frame
ready for ggplot2.
}
