% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP data}
\usage{
plotUMAP(plt_data)
}
\arguments{
\item{plt_data}{Data.frame of UMAP coordinates and labels.}
}
\value{
A \code{ggplot} object. A scatter plot of the UMAP coordinates of each item
coloured by the labels.
}
\description{
Plots the output of \code{makeUMAPPlotData}.
}
