% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseUMAP.R
\name{visualiseUMAP}
\alias{visualiseUMAP}
\title{Visualise UMAP}
\usage{
visualiseUMAP(x, labels)
}
\arguments{
\item{x}{Data.frame to perform UMAP upon.}

\item{labels}{A vector of labels associated with \code{x}.}
}
\value{
A \code{ggplot} object. A scatter plot of the UMAP coordinates of each item
coloured by the labels.
}
\description{
Wrapper function, creating and visualising UMAP
}
